start refactoring buildsys
