+++ packages/docker-build/print-info.sh
@@ -3,11 +3,8 @@ echo '==> $ /opt/devkitpro/devkitA64/bin/aarch64-none-elf-gcc --version'
 /opt/devkitpro/devkitA64/bin/aarch64-none-elf-gcc --version
 echo '==> $ gcc --version'
 gcc --version
-echo '==> $ megaton --version'
-megaton --version
-echo ''
-echo '==> $ megaton toolchain check'
-megaton toolchain check
+echo '==> $ megaton version -v'
+megaton version -v
 echo ''
 echo '==> $ rustup toolchain list'
 rustup toolchain list
